pyproject.toml
[project]
name = "osi-dump"
version = "0.1"
description = "OpenStack instances information dump tool"
readme = "README.md"
authors = [{ name = "TVKain" }]
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
classifiers = ["License :: OSI Approved :: Apache Software License"]
dependencies = ["typer", "openstacksdk", "pydantic", "pandas", "openpyxl"]


[project.scripts]
osi-dump = "osi_dump.cli:app"

[project.optional-dependencies]
test = ["pytest"]
